some cleanup after supporting 128-bit types
